Change how git-completion and git-prompt are applied

+++ install.sh
@@ -1,7 +1,11 @@
 #!/bin/sh
 
-ln -s -T /usr/share/bash-completion/completions/git ~/.git-completion.bash
-ln -s -T /usr/lib/git-core/git-sh-prompt ~/.git-prompt.sh
+if [ ! -L ~/.git-completion.bash ]; then
+    ln -s -T /usr/share/bash-completion/completions/git ~/.git-completion.bash
+fi
+if [ ! -L ~/.git-prompt.sh ]; then
+    ln -s -T /usr/lib/git-core/git-sh-prompt ~/.git-prompt.sh
+fi
 cp -T --backup=numbered --update bash_aliases ~/.bash_aliases
 cat bashrc.append >> ~/.bashrc
 cp -T --backup=numbered --update gitconfig ~/.gitconfig
